Add System:Remark to NestingSupported list
I was surprised by the fact that the "system:remark" is not in the nesting list but "system:remark with the event" is in.
Also I was surprized that Nano X supports adding "system:remark" into a batch of calls, but the default Nano S app version doesn't
We released mechanics which use this "system:remark" in a batch of calls because we tested it on Nano X and it's worked fine. Now we have to recommend our users installing the Polkadot XL app to use our mechanic.